# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Quotation Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /quotations:
    get:
      operationId: listQuotations
      tags: [quotations]
      summary: List quotations
      parameters:
        - { name: status, in: query, schema: { type: string } }
        - { name: clientId, in: query, schema: { type: string } }
      responses:
        "200":
          description: List of quotations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuotationSummary"
    post:
      operationId: createQuotation
      tags: [quotations]
      summary: Create a quotation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuotationInput"
      responses:
        "201":
          description: Created quotation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuotationDetail"
  /quotations/{id}:
    get:
      operationId: getQuotation
      tags: [quotations]
      summary: Get quotation by ID (with line items)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Quotation detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuotationDetail"
    put:
      operationId: updateQuotation
      tags: [quotations]
      summary: Update quotation
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuotationInput"
      responses:
        "200":
          description: Updated quotation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuotationDetail"
    delete:
      operationId: deleteQuotation
      tags: [quotations]
      summary: Delete quotation
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "204":
          description: Deleted
  /quotations/{id}/status:
    patch:
      operationId: changeQuotationStatus
      tags: [quotations]
      summary: Change quotation status
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [DRAFT, SENT, ACCEPTED, REJECTED, PAID, EXPIRED]
              required: [status]
      responses:
        "200":
          description: Updated quotation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuotationDetail"
  /quotations/{id}/duplicate:
    post:
      operationId: duplicateQuotation
      tags: [quotations]
      summary: Duplicate a quotation
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "201":
          description: New draft quotation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuotationDetail"
  /quotations/{id}/pdf:
    get:
      operationId: getQuotationPdf
      tags: [quotations]
      summary: Download quotation PDF
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: PDF binary
          content:
            application/pdf:
              schema:
                type: string
                format: binary
  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get company settings
      responses:
        "200":
          description: Company settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanySettings"
    put:
      operationId: updateSettings
      tags: [settings]
      summary: Upsert company settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingsInput"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanySettings"
  /settings/logo:
    post:
      operationId: uploadLogo
      tags: [settings]
      summary: Upload company logo
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                logo:
                  type: string
                  format: binary
      responses:
        "200":
          description: Uploaded logo URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  url: { type: string }
                  objectPath: { type: string }
...
components:
  schemas:
...
    Client:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        company: { type: string, nullable: true }
        email: { type: string }
        phone: { type: string, nullable: true }
        addressLine1: { type: string, nullable: true }
        addressLine2: { type: string, nullable: true }
        city: { type: string, nullable: true }
        region: { type: string, nullable: true }
        postalCode: { type: string, nullable: true }
        country: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, name, email, createdAt, updatedAt]
...
    LineItem:
      type: object
      properties:
        id: { type: string }
        quotationId: { type: string }
        description: { type: string }
        quantity: { type: string }
        unit: { type: string }
        unitPrice: { type: string }
        lineTotal: { type: string }
        position: { type: integer }
      required: [id, quotationId, description, quantity, unit, unitPrice, lineTotal, position]
    LineItemInput:
      type: object
      properties:
        description: { type: string }
        quantity: { type: number }
        unit: { type: string, enum: [hours, days, fixed] }
        unitPrice: { type: number }
        position: { type: integer }
      required: [description, quantity, unitPrice]
    QuotationInput:
      type: object
      properties:
        clientId: { type: string }
        issueDate: { type: string, format: date-time }
        validUntil: { type: string, format: date-time }
        currency: { type: string }
        discountType:
          type: string
          enum: [PERCENTAGE, FIXED]
          nullable: true
        discountValue: { type: number }
        taxRate: { type: number }
        notes: { type: string, nullable: true }
        terms: { type: string, nullable: true }
        paymentUrl: { type: string, nullable: true }
        showQrCode: { type: boolean }
        template: { type: string, enum: [MODERN, CLASSIC] }
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/LineItemInput"
      required: [clientId, issueDate, validUntil, lineItems]
    QuotationSummary:
      type: object
      properties:
        id: { type: string }
        number: { type: string }
        status: { type: string }
        total: { type: string }
        currency: { type: string }
        issueDate: { type: string, format: date-time }
        validUntil: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
        clientId: { type: string }
        clientName: { type: string, nullable: true }
        clientCompany: { type: string, nullable: true }
        template: { type: string }
        discountType: { type: string, nullable: true }
        discountValue: { type: string }
        taxRate: { type: string }
        subtotal: { type: string }
        discountAmount: { type: string }
        taxAmount: { type: string }
      required: [id, number, status, total, currency, issueDate, validUntil, clientId]
    QuotationDetail:
      allOf:
        - $ref: "#/components/schemas/QuotationSummary"
        - type: object
          properties:
            notes: { type: string, nullable: true }
            terms: { type: string, nullable: true }
            paymentUrl: { type: string, nullable: true }
            showQrCode: { type: boolean }
            sentAt: { type: string, format: date-time, nullable: true }
            acceptedAt: { type: string, format: date-time, nullable: true }
            paidAt: { type: string, format: date-time, nullable: true }
            client:
              $ref: "#/components/schemas/Client"
            lineItems:
              type: array
              items:
                $ref: "#/components/schemas/LineItem"
    CompanySettings:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        addressLine1: { type: string }
        addressLine2: { type: string, nullable: true }
        city: { type: string }
        region: { type: string, nullable: true }
        postalCode: { type: string }
        country: { type: string }
        phone: { type: string, nullable: true }
        email: { type: string }
        website: { type: string, nullable: true }
        taxNumber: { type: string, nullable: true }
        logoUrl: { type: string, nullable: true }
        currency: { type: string }
        defaultTaxRate: { type: string }
        defaultTerms: { type: string, nullable: true }
        defaultNotes: { type: string, nullable: true }
        defaultTemplate: { type: string }
        defaultPaymentUrl: { type: string, nullable: true }
        updatedAt: { type: string, format: date-time }
      required: [id, name, addressLine1, city, postalCode, country, email, currency, defaultTaxRate, defaultTemplate, updatedAt]
    SettingsInput:
      type: object
      properties:
        name: { type: string }
        addressLine1: { type: string }
        addressLine2: { type: string, nullable: true }
        city: { type: string }
        region: { type: string, nullable: true }
        postalCode: { type: string }
        country: { type: string }
        phone: { type: string, nullable: true }
        email: { type: string }
        website: { type: string, nullable: true }
        taxNumber: { type: string, nullable: true }
        logoUrl: { type: string, nullable: true }
        currency: { type: string }
        defaultTaxRate: { type: number }
        defaultTerms: { type: string, nullable: true }
        defaultNotes: { type: string, nullable: true }
        defaultTemplate: { type: string, enum: [MODERN, CLASSIC] }
        defaultPaymentUrl: { type: string, nullable: true }
      required: [name, addressLine1, city, postalCode, country, email, currency, defaultTaxRate, defaultTemplate]
...